WEB-2039 extract web part
